Add BaseExtractor base class and address PR comments
