+++ app/activities/extract_pdf_content.py
@@ -1,6 +1,3 @@
-from io import BytesIO
-from typing import Union
-
 import httpx
 from pydantic import BaseModel
 from temporalio import activity
@@ -12,8 +9,8 @@ class ExtractPdfContentRequest(BaseModel):
     """Request to extract PDF content from a URL."""
 
     url: str
-    extractor_type: str = "pdfplumber"  # Default to pdfplumber
-    pages: str | None = None            # Optional page specification such as "1,3-5,-1"
+    extractor: str = "pdfplumber"  # Default to pdfplumber
+    pages: list[int] | None = None  # Optional list of page numbers (1-indexed)
 
 
 class ExtractPdfContentResponse(BaseModel):
@@ -27,19 +24,19 @@ class ExtractPdfContentResponse(BaseModel):
 
 @activity.defn
 async def create(request: ExtractPdfContentRequest) -> ExtractPdfContentResponse:
-    """Download a PDF from a URL and extract its content using the specified extractor."""
+    """Download PDF from a URL and extract its content using the specified extractor."""
     async with httpx.AsyncClient() as client:
         response = await client.get(request.url)
         response.raise_for_status()
         pdf_bytes = response.content
 
     # Extract content using the extraction module
-    extractor = get_extractor(request.extractor_type)
+    extractor = get_extractor(request.extractor)
     result = extractor.extract(pdf_bytes, request.pages)
-    
+
     return ExtractPdfContentResponse(
         text=result["full_text"],
         num_pages=result["page_count"],
         pages_extracted=result["pages_extracted"],
-        extractor_used=request.extractor_type,
+        extractor_used=request.extractor,
     )
